# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Automated Code Compliance Checking System",
    "description": "Evidence-backed API foundation for automated building code compliance review.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/v1/access/api-keys": {
      "post": {
        "tags": [
          "release-candidate"
        ],
        "summary": "Create Api Key",
        "operationId": "create_api_key_api_v1_access_api_keys_post",
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "release-candidate"
        ],
        "summary": "List Api Keys",
        "operationId": "list_api_keys_api_v1_access_api_keys_get",
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeySummary"
                  },
                  "title": "Response List Api Keys Api V1 Access Api Keys Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/access/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "release-candidate"
        ],
        "summary": "Revoke Api Key",
        "operationId": "revoke_api_key_api_v1_access_api_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Key Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/pilot-feedback": {
      "post": {
        "tags": [
          "release-candidate"
        ],
        "summary": "Create Pilot Feedback",
        "operationId": "create_pilot_feedback_api_v1_projects__project_id__pilot_feedback_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotFeedbackCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "release-candidate"
        ],
        "summary": "List Pilot Feedback",
        "operationId": "list_pilot_feedback_api_v1_projects__project_id__pilot_feedback_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PilotFeedbackResponse"
                  },
                  "title": "Response List Pilot Feedback Api V1 Projects  Project Id  Pilot Feedback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pilot-feedback/{feedback_id}": {
      "patch": {
        "tags": [
          "release-candidate"
        ],
        "summary": "Update Pilot Feedback",
        "operationId": "update_pilot_feedback_api_v1_pilot_feedback__feedback_id__patch",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Feedback Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotFeedbackUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/overview": {
      "get": {
        "tags": [
          "release-candidate"
        ],
        "summary": "Operations Overview",
        "operationId": "operations_overview_api_v1_operations_overview_get",
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/audit-events": {
      "get": {
        "tags": [
          "release-candidate"
        ],
        "summary": "List Audit Events",
        "operationId": "list_audit_events_api_v1_operations_audit_events_get",
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "title": "Response List Audit Events Api V1 Operations Audit Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApiKeyCreate"
      },
      "ApiKeyCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "token",
          "expires_at",
          "created_at"
        ],
        "title": "ApiKeyCreated"
      },
      "ApiKeySummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "expires_at",
          "revoked_at",
          "last_used_at",
          "created_at"
        ],
        "title": "ApiKeySummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationsOverview": {
        "properties": {
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "jobs_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Jobs By Status"
          },
          "jobs_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Jobs By Type"
          },
          "check_runs": {
            "type": "integer",
            "title": "Check Runs"
          },
          "open_findings": {
            "type": "integer",
            "title": "Open Findings"
          },
          "stored_file_bytes": {
            "type": "integer",
            "title": "Stored File Bytes"
          },
          "pilot_feedback_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Pilot Feedback By Status"
          },
          "external_model_enabled": {
            "type": "boolean",
            "title": "External Model Enabled"
          },
          "external_model_calls": {
            "type": "integer",
            "title": "External Model Calls"
          },
          "estimated_model_cost": {
            "type": "number",
            "title": "Estimated Model Cost"
          },
          "deterministic_checks_available": {
            "type": "boolean",
            "title": "Deterministic Checks Available"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "organization_id",
          "jobs_by_status",
          "jobs_by_type",
          "check_runs",
          "open_findings",
          "stored_file_bytes",
          "pilot_feedback_by_status",
          "external_model_enabled",
          "external_model_calls",
          "estimated_model_cost",
          "deterministic_checks_available"
        ],
        "title": "OperationsOverview"
      },
...
      "PilotFeedbackCreate": {
        "properties": {
          "category": {
            "type": "string",
            "pattern": "^(false_positive|false_negative|evidence|usability|value|other)$",
            "title": "Category"
          },
          "severity": {
            "type": "string",
            "pattern": "^(low|medium|high|critical)$",
            "title": "Severity"
          },
          "summary": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Summary"
          },
          "details": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Details"
          },
          "time_saved_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Saved Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "severity",
          "summary",
          "details"
        ],
        "title": "PilotFeedbackCreate"
      },
      "PilotFeedbackResponse": {
        "properties": {
          "category": {
            "type": "string",
            "pattern": "^(false_positive|false_negative|evidence|usability|value|other)$",
            "title": "Category"
          },
          "severity": {
            "type": "string",
            "pattern": "^(low|medium|high|critical)$",
            "title": "Severity"
          },
          "summary": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Summary"
          },
          "details": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Details"
          },
          "time_saved_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Saved Minutes"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "submitted_by_id": {
            "type": "string",
            "format": "uuid",
            "title": "Submitted By Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "severity",
          "summary",
          "details",
          "id",
          "project_id",
          "submitted_by_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "PilotFeedbackResponse"
      },
      "PilotFeedbackUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(open|triaged|resolved|accepted_limit)$",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PilotFeedbackUpdate"
      },
...
    }
  }
}